Adjusting training set to be more realistic
I intentionally made the training set 'aggressive', in the
hopes that that would allow the model to deal with complex
situations.

However, the model has become underconfident in simple cases.
This commit is an attempt to see if that's a matter of too
complex training (too much curl, too many fibres).